비밀번호 변경 추가. 기존 재설정 페이지에서 분기처리

+++ src/main/java/com/imchobo/sayren_back/domain/member/service/MemberServiceImpl.java
@@ -137,13 +137,21 @@ public class MemberServiceImpl implements MemberService {
   @Override
   @Transactional
   public void changePassword(ResetPasswordRequestDTO resetPasswordRequestDTO) {
-    Long memberId = redisUtil.getResetPassword(resetPasswordRequestDTO.getToken());
-    redisUtil.deleteResetPassword(resetPasswordRequestDTO.getToken());
+    Long memberId;
+    if(SecurityUtil.isUser()){
+      memberId = SecurityUtil.getMemberAuthDTO().getId();
+    }
+    else {
+      memberId = redisUtil.getResetPassword(resetPasswordRequestDTO.getToken());
+    }
     Member member = memberRepository.findById(memberId).orElseThrow(IllegalArgumentException::new);
     if(passwordEncoder.matches(resetPasswordRequestDTO.getNewPassword(), member.getPassword())) {
       throw new PasswordAlreadyUseException();
     }
     member.setPassword(passwordEncoder.encode(resetPasswordRequestDTO.getNewPassword()));
+    if(resetPasswordRequestDTO.getToken() != null) {
+      redisUtil.deleteResetPassword(resetPasswordRequestDTO.getToken());
+    }
   }
 
 
@@ -176,5 +184,14 @@ public class MemberServiceImpl implements MemberService {
     }
     member.setName(changeNameDTO.getName());
     return memberMapper.toLoginResponseDTO(memberMapper.toAuthDTO(member));
+  }
+
+
+  @Override
+  public void passwordCheck(PasswordCheckDTO passwordCheckDTO) {
+    Member member = memberRepository.findById(SecurityUtil.getMemberAuthDTO().getId()).orElseThrow(IllegalArgumentException::new);
+    if (!passwordEncoder.matches(passwordCheckDTO.getPassword(), member.getPassword())) {
+      throw new InvalidPasswordException();
+    }
   }
 }
